Fix flow and show mint page
