+++ src/components/Mint.module.css
@@ -27,17 +27,6 @@
 
 /* TASK: rewrite the above styled component with CSS */
 
-.StyledContainer {
-  width: 100vw;
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 10000;
-}
-
-/* TASK: rewrite the above styled component with CSS */
-
 .Title {
   display: flex;
   justify-content: center;
@@ -54,20 +43,12 @@
 
 /* TASK: rewrite the above styled component with CSS */
 
-.StyledBackground {
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  background-color: #000000;
-}
-
-/* TASK: rewrite the above styled component with CSS */
-
-.StyledPopup {
+.StyledComponent {
   width: 550px;
   display: flex;
   flex-direction: column;
-  position: relative;
+  position: fixed;
+  bottom: 0;
   background-color: #1716168d;
   border-color: #38404e;
   border-style: solid;
