+++ src/pages/Create.module.css
@@ -1,0 +1,4 @@
+.classdisabled {
+    pointer-events: none;
+    opacity: 0.5;
+}
